Decrease craigslist testdata from 2.8M to 20K and movies from 110K to 50K. Most test cases load these files and other every single run, which leads to them spending most of their time just Unmarshaling and Marshaling data. Keep the old, huge testdata around in the repo/test/test_huge_dataset directory, because they're probably good for running at least 1 test on, to make sure our code can handle relatively large datasets.

Performance numbers from before this change, taken from https://github.com/qri-io/qri/pull/527:
```
Run 0
ok      github.com/qri-io/qri/api      13.866s
ok      github.com/qri-io/qri/cmd      61.417s
ok      github.com/qri-io/qri/config   32.168s
ok      github.com/qri-io/qri/lib      32.685s
ok      github.com/qri-io/qri/p2p       8.008s
Run 1
ok      github.com/qri-io/qri/api      18.329s
ok      github.com/qri-io/qri/cmd      68.314s
ok      github.com/qri-io/qri/config   43.116s
ok      github.com/qri-io/qri/lib      32.584s
ok      github.com/qri-io/qri/p2p       6.926s
```

After this change:
```
Run 0:
ok      github.com/qri-io/qri/api       10.498s
ok      github.com/qri-io/qri/cmd       54.005s
ok      github.com/qri-io/qri/config    36.107s
ok      github.com/qri-io/qri/lib       14.220s
ok      github.com/qri-io/qri/p2p       4.917s
Run 1:
ok      github.com/qri-io/qri/api       11.043s
ok      github.com/qri-io/qri/cmd       53.763s
ok      github.com/qri-io/qri/config    52.061s
ok      github.com/qri-io/qri/lib       18.909s
ok      github.com/qri-io/qri/p2p       6.475s
```

`api` speeds up significantly, and `lib` is nearly cut in half. No idea what's going on with `config`, though, it has a huge variance.